Move hasLocalCodeInAlternateCoding to HapiHelper for reuse
